TASK: Allow node migrations only to run on workspaces based on root workspaces (e.g. migration-ws based on live)
This enables us to do meaningful constraint checks versus changes in other workspaces.

https://neos-project.slack.com/archives/C04PYL8H3/p1741976899524389
