+++ Neos.ContentRepository.Core/Classes/Feature/Common/WorkspaceConstraintChecks.php
@@ -1,0 +1,44 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Neos\ContentRepository\Core\Feature\Common;
+
+use Neos\ContentRepository\Core\CommandHandler\CommandHandlingDependencies;
+use Neos\ContentRepository\Core\Feature\WorkspaceCreation\Exception\BaseWorkspaceDoesNotExist;
+use Neos\ContentRepository\Core\SharedModel\Exception\WorkspaceDoesNotExist;
+use Neos\ContentRepository\Core\SharedModel\Exception\WorkspaceHasNoBaseWorkspaceName;
+use Neos\ContentRepository\Core\SharedModel\Workspace\Workspace;
+use Neos\ContentRepository\Core\SharedModel\Workspace\WorkspaceName;
+
+trait WorkspaceConstraintChecks
+{
+    /**
+     * @throws WorkspaceDoesNotExist
+     */
+    private function requireWorkspace(WorkspaceName $workspaceName, CommandHandlingDependencies $commandHandlingDependencies): Workspace
+    {
+        $workspace = $commandHandlingDependencies->findWorkspaceByName($workspaceName);
+        if (is_null($workspace)) {
+            throw WorkspaceDoesNotExist::butWasSupposedTo($workspaceName);
+        }
+
+        return $workspace;
+    }
+
+    /**
+     * @throws WorkspaceHasNoBaseWorkspaceName
+     * @throws BaseWorkspaceDoesNotExist
+     */
+    private function requireBaseWorkspace(Workspace $workspace, CommandHandlingDependencies $commandHandlingDependencies): Workspace
+    {
+        if (is_null($workspace->baseWorkspaceName)) {
+            throw WorkspaceHasNoBaseWorkspaceName::butWasSupposedTo($workspace->workspaceName);
+        }
+        $baseWorkspace = $commandHandlingDependencies->findWorkspaceByName($workspace->baseWorkspaceName);
+        if (is_null($baseWorkspace)) {
+            throw BaseWorkspaceDoesNotExist::butWasSupposedTo($workspace->workspaceName);
+        }
+        return $baseWorkspace;
+    }
+}
